iOS bundle build action: Build bundle without code signing.
Code signing in CI is problematic because there is no legitimate signing
identity to sign the code. Since we're only intersted in the iOS bundle
and not in running the InjectionIII app, it's ok to drop code signing.

In addition, since the `xcodebuild` steps in `build_bundles.sh` doesn't
take into account the `derivedDataPath` passed to main `xcodebuild`, the
bundles won't be written to `output` but to the default derived data
directory. Therefore the `derivedDataPath` is dropped and the iOS bundle
is copied from the default derived data directory instead.