build_bundles: Build for iPhone instead of AppleTV.
In order for this to work, `CODE_SIGNING_ALLOWED` is set to NO. The iOS
bundle will be signed by the app that uses it.

`PRODUCT_MODULE_NAME` is used in `build_bundle` because there are files
that expect the generated Swift file to have the name
`iOSInjection-Swift.h`, which takes its name from the `PRODUCT_NAME`,
unless `PRODUCT_MODULE_NAME` is specified.

Also, remove redundant copy to `output` folder.